Move the code to cli.py
